Create CRUD Operations in Services Directory
I created the file BookService.cs which has CRUD operations for the
BookstoreDB using MongoDB.  I also added a connection to the DB in
appsettings.json file.  This is needed in order for the application to
connect to the correct database before operations can be performed.

I need to register the BookService class with the dependency injection
system in the Startup file, otherwise the application will not recognize
the operations inputted by the user.